Fix logout (probably)
Currently, there is a race condition when clicking the logout link:

1. JavaScript gets executed, initiating XHRs
A2. The link is followed, initiating a navigation
A3. Upon navigation, all pending requests are cancelled
A4. / (the frontpage) is rendered.
B2. The XHR gets sent to the server (which starts sending a response)
B3. Upon receiving the XHR, dbas initiates window.reload()

a) If B3 happens before A3, everything works as intended.
b) If B2 happens before A3, but B3 after A3, we end up at the frontpage (/), but logged out.
c) If A3 completes before B2, B will be cancelled; we end up at the frontpage still logged in.

If the webbrowser has a connection already running and very quickly sends its data,  b) will happen.
However, it is perfectly possible that the browser does not have a connection already and needs to connect() again. In that case, c) will happen.
a) can happen when dbas is run locally.

To fix this, prevent the default navigation action of links (maybe they shouldn't be HTML links in the first place?). This will be necessary for a lot of other links as well.

While we're at it, also remove the ajaxComplete call - ajaxComplete is a per-window handler, not a per-request one, so even if it were available it wouldn't work here.