Correctly expect a statement id instead of a premisegroup id
This workaround only works if the there is only one premisegroup for the
statement. The better solution would be actually passing the
premisegroup id in the frontend (which can only handle statement and
argument ids at the moment). Before that, we must discuss whether we
want premises or statements to be split.